complyAI_FT_scripts/requirements.txt
datasets==2.16.1
PyPDF2==3.0.1
transformers
langchain
python-dotenv==1.0.1
langchain-mistralai
langchain-huggingface
langchain_community
langchain-experimental
pyyaml==6.0.1
mistral-common>=1.3.1
mdc==1.2.1
pandas
mistralai